backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
celery==5.3.4
redis==5.0.1
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
schedule==1.2.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1 